resources: ROCm 7.2
Bump ROCm and pytorch to 7.2. This disk image requires new gem5 changes
to implement better ACPI support. The gem5_wmi workaround module is
removed in favor of this due to the workaround no longer being possible
in latest kernel versions. This version also adds the gem5-bridge
symlink and contains the gem5_init.sh script to use it.

+++ src/x86-ubuntu-gpu-ml/scripts/rocm-install.sh
@@ -3,9 +3,12 @@
 # Copyright (c) 2024 The Regents of the University of California.
 # SPDX-License-Identifier: BSD 3-Clause
 
-# Copyright (c) 2024 Advanced Micro Devices, Inc.
-# All rights reserved.
+# Copyright (c) 2024-2026 Advanced Micro Devices, Inc.
 # SPDX-License-Identifier: BSD 3-Clause
+
+# Install a known-working version of Linux as this might change after stable
+# release.
+KERNEL=6.17.0-35-generic
 
 # Installing the packages in this script instead of the user-data
 # file dueing ubuntu autoinstall. The reason is that sometimes
@@ -17,39 +20,18 @@ apt-get install -y git
 apt-get install -y vim
 apt-get install -y build-essential
 
+# Make sure the init script is executable
+chmod a+x /home/gem5/gem5_init.sh
+mv /home/gem5/gem5_init.sh /sbin/
+mv /sbin/init /sbin/init.gem5.bak
+ln -s /sbin/gem5_init.sh /sbin/init
+
 # Remove the motd
 rm /etc/update-motd.d/*
 
-# Build the m5 util
-git clone https://github.com/gem5/gem5.git --depth=1 --filter=blob:none --no-checkout --sparse --single-branch --branch=stable
-pushd gem5
-# Checkout just the files we need
-git sparse-checkout add util/m5
-git sparse-checkout add include
-git checkout
-# Build the library and binary
-pushd util/m5
-scons build/x86/out/m5
-cp build/x86/out/m5 /sbin/m5
-popd
-popd
-rm -rf gem5
-
-# Occasionally connecting to github fails. Bail now instead of making a disk
-# image that is not usable.
-if [ ! -f /sbin/m5 ]; then
-    echo "m5 util did not appear to build correctly. This disk will not be usable."
-    echo "Try to build the disk again if there was a temporary error (e.g., not able to connect to github)."
-    echo "For other problems create an issue at https://github.com/gem5/gem5/issues."
-    exit 1
-fi
-
-# Make the script to load amdgpu module with the above module executable
-chmod a+x /home/gem5/load_amdgpu.sh
-
 # The following instructions were obtained from the ROCm installation guide:
-# https://rocm.docs.amd.com/projects/install-on-linux/en/latest/how-to/
-#     native-install/ubuntu.html
+# https://instinct.docs.amd.com/projects/amdgpu-docs/en/docs-30.30.4/install/
+#           detailed-install/package-manager/package-manager-ubuntu.html
 
 # Make the directory if it doesn't exist yet.
 # This location is recommended by the distribution maintainers.
@@ -60,16 +42,15 @@ sudo mkdir --parents --mode=0755 /etc/apt/keyrings
 wget https://repo.radeon.com/rocm/rocm.gpg.key -O - | \
         gpg --dearmor | sudo tee /etc/apt/keyrings/rocm.gpg > /dev/null
 
-echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/rocm.gpg] https://repo.radeon.com/amdgpu/7.0/ubuntu noble main" \
+echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/rocm.gpg] https://repo.radeon.com/amdgpu/7.2.4/ubuntu noble main" \
         | sudo tee /etc/apt/sources.list.d/amdgpu.list
 
-echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/rocm.gpg] https://repo.radeon.com/rocm/apt/7.0 noble main" \
+echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/rocm.gpg] https://repo.radeon.com/rocm/apt/7.2.4 noble main" \
         | sudo tee --append /etc/apt/sources.list.d/rocm.list
 echo -e 'Package: *\nPin: release o=repo.radeon.com\nPin-Priority: 600' \
         | sudo tee /etc/apt/preferences.d/rocm-pin-600
 sudo apt update
 
-sudo apt -y install amdgpu-dkms
 sudo apt -y install rocm
 sudo apt -y install cmake
 
@@ -85,25 +66,11 @@ sudo touch /usr/lib/firmware/amdgpu/ip_discovery.bin
 sudo chmod 777 /usr/lib/firmware/amdgpu/ip_discovery.bin
 
 
-# Install a known-working version of Linux as this might change after stable
-# release. Install this after DKMS so they are rebuilt.
-KERNEL=6.8.0-79-generic
-
 sudo apt -y install "linux-image-${KERNEL}"
 sudo apt -y install "linux-headers-${KERNEL}" "linux-modules-extra-${KERNEL}"
 
 echo "Extracting linux kernel"
 sudo bash -c "/usr/src/linux-headers-${KERNEL}/scripts/extract-vmlinux /boot/vmlinuz-${KERNEL} > /home/gem5/vmlinux-gpu-ml"
-
-# Build the gem5 Linux module containing symbols missing due to outdated ACPI support in gem5
-pushd /home/gem5
-make -C /lib/modules/${KERNEL}/build M=${PWD}
-if [ ! -f ./gem5_wmi.ko ]; then
-    echo "gem5_wmi module did not appear to build correctly. This disk will not be usable."
-    echo "Please post an issue at https://github.com/gem5/gem5/issues with this output log."
-    exit 1
-fi
-popd
 
 # Make the discovery files writeable by packer
 touch /usr/lib/firmware/amdgpu/mi300_discovery
@@ -112,25 +79,68 @@ touch /usr/lib/firmware/amdgpu/mi350_discovery
 chmod 777 /usr/lib/firmware/amdgpu/mi300_discovery
 chmod 777 /usr/lib/firmware/amdgpu/mi350_discovery
 
+# Allow internet access when running this disk e.g., in systemd-nspawn
+rm /etc/resolv.conf
+echo "nameserver 127.0.0.53" > /etc/resolv.conf
+
 # Note about pip: This disk is created for the express purpose of being run in
 # gem5 and is therefore effectively sandboxed enough that we can use the pip
 # option --break-system-packages. If you plan to modify this disk image with
 # pip packages that might conflict, it is up to you to resolve the conflicts.
 
 # See https://pytorch.org/.
-# Build: 2.3.0
+# Build: 2.12.0
 # OS: Linux
 # Package: Pip
 # Language: Python
-# Compute Platfrom: ROCm 6.2.4
+# Compute Platfrom: ROCm 7.2
 sudo apt -y install pip3
-pip3 install --break-system-packages torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.0
+pip3 install --break-system-packages torch torchvision --index-url https://download.pytorch.org/whl/rocm7.2
 
-# For a newer version uncomment one below and remove the above install:
-# Warning: Absurdly slow compared to ROCm 6.0 *in simulation*:
-#pip3 install --break-system-packages torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.4
-# Warning: nightly build, may not work depending on day. Use at your own risk:
-#pip3 install --pre torch torchvision --index-url https://download.pytorch.org/whl/nightly/rocm7.0 --break-system-packages
+# Try and import pytorch, exit failure if it does not exist
+python3 -c "import torch"
+if [ "$?" != "0" ]; then
+    echo "Pytorch installation failed. See log."
+    exit 1
+fi
+
+# Build the m5 util
+git clone https://github.com/gem5/gem5.git --depth=1 --filter=blob:none --no-checkout --sparse --single-branch --branch=stable
+pushd gem5
+# Checkout just the files we need
+git sparse-checkout add util/m5
+git sparse-checkout add util/gem5_bridge
+git sparse-checkout add include
+git checkout
+# Build the library and binary
+pushd util/m5
+scons build/x86/out/m5
+cp build/x86/out/m5 /usr/local/bin/
+cp build/x86/out/libm5.a /usr/local/lib/
+popd # util/m5
+pushd util/gem5_bridge
+make build install KVERSION=${KERNEL}
+depmod --quick
+popd
+popd
+
+if [ ! -f /usr/local/bin/m5 ]; then
+    echo "m5 util did not appear to build correctly. This disk will not be usable."
+    echo "Try to build the disk again if there was a temporary error (e.g., not able to connect to github)."
+    echo "For other problems create an issue at https://github.com/gem5/gem5/issues."
+    exit 1
+fi
+
+mv /usr/local/bin/m5 /usr/local/bin/gem5-bridge
+
+chmod 4755 /usr/local/bin/gem5-bridge
+chmod u+s /usr/local/bin/gem5-bridge
+
+ln -s /usr/local/bin/gem5-bridge /usr/local/bin/m5
+
+rm -rf gem5
+
+echo "gem5-bridge complete!"
 
 # Setup gem5 auto login.
 mv /home/gem5/serial-getty@.service /lib/systemd/system/
